Add validation for MangaPage Controller

+++ Manga/MediatR/MangaPage/Commands/Delete/DeletePageCommandHadnler.cs
@@ -1,4 +1,5 @@
-﻿using Manga.Services.IServices;
+﻿using Manga.Exceptions;
+using Manga.Services.IServices;
 using MediatR;
 
 namespace Manga.MediatR.MangaPage.Commands.Delete
@@ -14,6 +15,10 @@ namespace Manga.MediatR.MangaPage.Commands.Delete
         public async Task Handle(DeletePageCommand request, CancellationToken cancellationToken)
         {
             var page = await _pageService.GetByIdAsync(request.id);
+            if(page is null)
+            {
+                throw new NotFoundException($"There is no page with id {request.id}");
+            }
             await _pageService.RemoveAsync(page);
         }
     }
